Move new TUF autoupdater to its own package
