+++ pkg/make/builder.go
@@ -412,17 +412,20 @@ func (b *Builder) GenerateTUF(ctx context.Context) error {
 		}
 	}
 
+	if err := b.execBindata(ctx, "pkg/autoupdate/assets/..."); err != nil {
+		return fmt.Errorf("exec bindata for autoupdate assets: %w", err)
+	}
+
 	// Bootstrap new TUF also -- for now, using the dev TUF repo until the production one is available
-	localRepo := filepath.Join("pkg", "autoupdate", "assets", "tuf-dev")
+	localRepo := filepath.Join("pkg", "tuf", "assets", "tuf-dev")
 	if err := os.MkdirAll(localRepo, 0755); err != nil {
-		return fmt.Errorf("make autoupdate dir %s: %w", localRepo, err)
+		return fmt.Errorf("make TUF autoupdate dir %s: %w", localRepo, err)
 	}
 	if err := bootstrapFromTUF("https://tuf-devel.kolide.com", localRepo); err != nil {
 		return fmt.Errorf("bootstrap TUF: %w", err)
 	}
-
-	if err := b.execBindata(ctx, "pkg/autoupdate/assets/..."); err != nil {
-		return fmt.Errorf("exec bindata for autoupdate assets: %w", err)
+	if err := b.execBindata(ctx, "pkg/tuf/assets/..."); err != nil {
+		return fmt.Errorf("exec bindata for TUF autoupdate assets: %w", err)
 	}
 
 	return nil
